+++ android/src/main/kotlin/dev/applicazza/flutter/plugins/whatsapp_stickers/ConfigFileManager.java
@@ -33,21 +33,21 @@ public class ConfigFileManager {
 
     public static final String CONTENT_FILE_NAME = "sticker_packs.json";
 
-    static List<StickerPack> getStickerPacks(Context context){
+    static List<StickerPack> getStickerPacks(Context context) {
         List<StickerPack> stickerPackList;
         File file = new File(getConfigFilePath(context));
-        if(!file.exists()){
+        if (!file.exists()) {
             return new ArrayList<StickerPack>();
         }
         try (InputStream contentsInputStream = new FileInputStream(file)) {
-           stickerPackList = ContentFileParser.parseStickerPacks(contentsInputStream);
+            stickerPackList = ContentFileParser.parseStickerPacks(contentsInputStream);
         } catch (IOException | IllegalStateException e) {
-            throw new RuntimeException("config file has some issues: " + e.getMessage(), e);
+            throw new RuntimeException("config file has some issues hahaha: " + e.getMessage(), e);
         }
         return stickerPackList;
     }
 
-    static StickerPack fromMethodCall(Context context, MethodCall call){
+    static StickerPack fromMethodCall(Context context, MethodCall call) {
         String identifier = call.argument("identifier");
         String name = call.argument("name");
         String publisher = call.argument("publisher");
@@ -56,13 +56,15 @@ public class ConfigFileManager {
         String publisherWebsite = call.argument("publisherWebsite");
         String privacyPolicyWebsite = call.argument("privacyPolicyWebsite");
         String licenseAgreementWebsite = call.argument("licenseAgreementWebsite");
+        String imageDataVersion = call.argument("imageDataVersion");
         Map<String, List<String>> stickers = call.argument("stickers");
-
-        StickerPack newStickerPack = new StickerPack(identifier, name, publisher, trayImageFileName, "", publisherWebsite, privacyPolicyWebsite, licenseAgreementWebsite, "1", false);
+        StickerPack newStickerPack = new StickerPack(identifier, name, publisher, trayImageFileName, "",
+                publisherWebsite, privacyPolicyWebsite, licenseAgreementWebsite, imageDataVersion, false);
         List<Sticker> newStickers = new ArrayList<Sticker>();
         assert stickers != null;
         for (Map.Entry<String, List<String>> entry : stickers.entrySet()) {
             Sticker s = new Sticker(getFileName(entry.getKey()), entry.getValue());
+
             newStickers.add(s);
         }
         newStickerPack.setStickers(newStickers);
@@ -73,37 +75,38 @@ public class ConfigFileManager {
 
     static boolean addNewPack(Context context, StickerPack stickerPack) throws JSONException, InvalidPackException {
         List<StickerPack> stickerPacks = new ArrayList<StickerPack>();
-        for(StickerPack s: getStickerPacks(context)){
-            if(!s.identifier.equals(stickerPack.identifier)){
+        for (StickerPack s : getStickerPacks(context)) {
+            if (!s.identifier.equals(stickerPack.identifier)) {
                 stickerPacks.add(s);
             }
         }
         stickerPacks.add(stickerPack);
-        return updateConfigFile(context,stickerPacks);
+        return updateConfigFile(context, stickerPacks);
     }
 
-    static String getFileName(String name){
-        if(name.contains("assets://")){
+    static String getFileName(String name) {
+        if (name.contains("assets://")) {
             name = name.replace("assets://", "");
-            name = name.replace("/", "_MZN_AD_");
-        }else if(name.contains("file://")) {
+            name = name.replace("/", "mzn_ad_");
+        } else if (name.contains("file://")) {
             name = name.replace("file://", "");
-            name = name.replace("/", "_MZN_FD_");
+            name = name.replace("/", "mzn_fd_");
         }
         return name;
     }
 
-    static boolean updateConfigFile(Context context, List<StickerPack> stickerPacks) throws JSONException, InvalidPackException {
+    static boolean updateConfigFile(Context context, List<StickerPack> stickerPacks)
+            throws JSONException, InvalidPackException {
         JSONObject mObj = new JSONObject();
-        if(stickerPacks.size() <= 0){
+        if (stickerPacks.size() <= 0) {
             mObj.put("android_play_store_link", "");
             mObj.put("ios_app_store_link", "");
-        }else{
+        } else {
             mObj.put("android_play_store_link", stickerPacks.get(0).androidPlayStoreLink);
             mObj.put("ios_app_store_link", stickerPacks.get(0).iosAppStoreLink);
         }
         JSONArray _packs = new JSONArray();
-        for(StickerPack s: stickerPacks){
+        for (StickerPack s : stickerPacks) {
             JSONObject obj = new JSONObject();
             obj.put("identifier", s.identifier);
             obj.put("name", s.name);
@@ -117,12 +120,12 @@ public class ConfigFileManager {
             obj.put("license_agreement_website", s.licenseAgreementWebsite);
 
             JSONArray stickerList = new JSONArray();
-            for (Sticker _sticker: s.getStickers()) {
+            for (Sticker _sticker : s.getStickers()) {
                 JSONObject stickerObj = new JSONObject();
                 String stickerFileName = getFileName(_sticker.imageFileName);
                 stickerObj.put("image_file", stickerFileName);
                 JSONArray _emojies = new JSONArray();
-                for(String emoji: _sticker.emojis){
+                for (String emoji : _sticker.emojis) {
                     _emojies.put(emoji);
                 }
                 stickerObj.put("emojis", _emojies);
@@ -136,7 +139,7 @@ public class ConfigFileManager {
         return true;
     }
 
-    static void writeConfigFile(Context context, String jsonString){
+    static void writeConfigFile(Context context, String jsonString) {
         String filePath = getConfigFilePath(context);
         File f = new File(filePath);
         try {
@@ -150,8 +153,8 @@ public class ConfigFileManager {
 
     static void generateConfigFile(Context context) throws JSONException {
         File file = new File(getConfigFilePath(context));
-        if(!file.exists()){
-            //prepare data
+        if (!file.exists()) {
+            // prepare data
             JSONObject mObj = new JSONObject();
             mObj.put("android_play_store_link", "");
             mObj.put("ios_app_store_link", "");
@@ -162,7 +165,7 @@ public class ConfigFileManager {
         }
     }
 
-    public static String getConfigFilePath(Context context){
+    public static String getConfigFilePath(Context context) {
         return PathUtils.getDataDirectory(context) + File.separator + CONTENT_FILE_NAME;
     }
 }