+++ android/src/main/kotlin/dev/applicazza/flutter/plugins/whatsapp_stickers/WhatsappStickersPlugin.kt
@@ -136,6 +136,7 @@ public class WhatsappStickersPlugin: FlutterPlugin, MethodCallHandler, ActivityA
   }
 
   override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+    this.activity = null
     channel.setMethodCallHandler(null)
   }
 
@@ -161,7 +162,6 @@ public class WhatsappStickersPlugin: FlutterPlugin, MethodCallHandler, ActivityA
   }
 
   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?): Boolean {
-
     if (requestCode == ADD_PACK) {
       // Create the intent
       if (resultCode == Activity.RESULT_CANCELED) {
@@ -169,6 +169,8 @@ public class WhatsappStickersPlugin: FlutterPlugin, MethodCallHandler, ActivityA
           val validationError = data.getStringExtra("validation_error")
           if (validationError != null) {
             this.result?.error("error", validationError, "")
+          } else {
+            this.result?.error("cancelled", "cancelled", "")
           }
         } else {
           this.result?.error("cancelled", "cancelled", "")