+++ lib/exceptions.dart
@@ -2,84 +2,122 @@ class WhatsappStickersException implements Exception {
   final String? cause;
 
   WhatsappStickersException(this.cause);
+
+  @override
+  String toString() => 'Unknown error: $cause';
 }
 
 class WhatsappStickersFileNotFoundException extends WhatsappStickersException {
-  static const String CODE = 'FILE_NOT_FOUND';
+  static const String code = 'FILE_NOT_FOUND';
 
-  final String? cause;
+  WhatsappStickersFileNotFoundException(String? cause) : super(cause);
 
-  WhatsappStickersFileNotFoundException(this.cause) : super('');
+  @override
+  String toString() => 'File not found';
 }
 
-class WhatsappStickersNumOutsideAllowableRangeException extends WhatsappStickersException {
-  static const String CODE = 'OUTSIDE_ALLOWABLE_RANGE';
+class WhatsappStickersNumOutsideAllowableRangeException
+    extends WhatsappStickersException {
+  static const String code = 'OUTSIDE_ALLOWABLE_RANGE';
 
-  final String? cause;
+  WhatsappStickersNumOutsideAllowableRangeException(String? cause)
+      : super(cause);
 
-  WhatsappStickersNumOutsideAllowableRangeException(this.cause) : super('');
+  @override
+  String toString() => 'Outside allowable range';
 }
 
-class WhatsappStickersUnsupportedImageFormatException extends WhatsappStickersException {
-  static const String CODE = 'UNSUPPORTED_IMAGE_FORMAT';
+class WhatsappStickersUnsupportedImageFormatException
+    extends WhatsappStickersException {
+  static const String code = 'UNSUPPORTED_IMAGE_FORMAT';
 
-  final String? cause;
+  WhatsappStickersUnsupportedImageFormatException(String? cause) : super(cause);
 
-  WhatsappStickersUnsupportedImageFormatException(this.cause) : super('');
+  @override
+  String toString() => 'Unsupported image format';
 }
 
 class WhatsappStickersImageTooBigException extends WhatsappStickersException {
-  static const String CODE = 'IMAGE_TOO_BIG';
+  static const String code = 'IMAGE_TOO_BIG';
 
-  final String? cause;
+  WhatsappStickersImageTooBigException(String? cause) : super(cause);
 
-  WhatsappStickersImageTooBigException(this.cause) : super('');
+  @override
+  String toString() => 'Image too big';
 }
 
-class WhatsappStickersIncorrectImageSizeException extends WhatsappStickersException {
-  static const String CODE = 'INCORRECT_IMAGE_SIZE';
+class WhatsappStickersIncorrectImageSizeException
+    extends WhatsappStickersException {
+  static const String code = 'INCORRECT_IMAGE_SIZE';
 
-  final String? cause;
+  WhatsappStickersIncorrectImageSizeException(String? cause) : super(cause);
 
-  WhatsappStickersIncorrectImageSizeException(this.cause) : super('');
+  @override
+  String toString() => 'Incorrect image size';
 }
 
-class WhatsappStickersAnimatedImagesNotSupportedException extends WhatsappStickersException {
-  static const String CODE = 'ANIMATED_IMAGES_NOT_SUPPORTED';
+class WhatsappStickersAnimatedImagesNotSupportedException
+    extends WhatsappStickersException {
+  static const String code = 'ANIMATED_IMAGES_NOT_SUPPORTED';
 
-  final String? cause;
+  WhatsappStickersAnimatedImagesNotSupportedException(String? cause)
+      : super(cause);
 
-  WhatsappStickersAnimatedImagesNotSupportedException(this.cause) : super('');
+  @override
+  String toString() => 'Animated images not supported';
 }
 
 class WhatsappStickersTooManyEmojisException extends WhatsappStickersException {
-  static const String CODE = 'TOO_MANY_EMOJIS';
+  static const String code = 'TOO_MANY_EMOJIS';
 
-  final String? cause;
+  WhatsappStickersTooManyEmojisException(String? cause) : super(cause);
 
-  WhatsappStickersTooManyEmojisException(this.cause) : super('');
+  @override
+  String toString() => 'Too many emojis';
 }
 
 class WhatsappStickersEmptyStringException extends WhatsappStickersException {
-  static const String CODE = 'EMPTY_STRING';
+  static const String code = 'EMPTY_STRING';
 
-  final String? cause;
+  WhatsappStickersEmptyStringException(String? cause) : super(cause);
 
-  WhatsappStickersEmptyStringException(this.cause) : super('');
+  @override
+  String toString() => 'Contain empty string';
 }
 
 class WhatsappStickersStringTooLongException extends WhatsappStickersException {
-  static const String CODE = 'STRING_TOO_LONG';
+  static const String code = 'STRING_TOO_LONG';
 
-  final String? cause;
+  WhatsappStickersStringTooLongException(String? cause) : super(cause);
 
-  WhatsappStickersStringTooLongException(this.cause) : super('');
+  @override
+  String toString() => 'String too long';
 }
 
-class WhatsappStickersIncorrectSizeTrayImageException extends WhatsappStickersException {
-  static const String CODE = 'INCORRECT_SIZE_TRAY_IMAGE';
+class WhatsappStickersAlreadyAddedException extends WhatsappStickersException {
+  static const String code = 'ALREADY_ADDED';
 
-  final String? cause;
+  WhatsappStickersAlreadyAddedException(String? cause) : super(cause);
 
-  WhatsappStickersIncorrectSizeTrayImageException(this.cause) : super('');
+  @override
+  String toString() => 'Stickers already added';
+}
+
+class WhatsappStickersCancelledException extends WhatsappStickersException {
+  static const String code = 'CANCELLED';
+
+  WhatsappStickersCancelledException(String? cause) : super(cause);
+
+  @override
+  String toString() => 'Cancelled';
+}
+
+class WhatsappStickersIncorrectSizeTrayImageException
+    extends WhatsappStickersException {
+  static const String code = 'INCORRECT_SIZE_TRAY_IMAGE';
+
+  WhatsappStickersIncorrectSizeTrayImageException(String? cause) : super(cause);
+
+  @override
+  String toString() => 'Incorrect size';
 }