+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.1.1
+
+* Overwrite `toString()` for exceptions
+
+## 1.1.0
+
+* Added 2 extra exceptions `WhatsappStickersAlreadyAddedException` and `WhatsappStickersCancelledException`
+* Fixed missing else statement handling on Android when cancel prompt
+
+## 1.0.0
+
+* Upgraded dependencies to make sure Android and iOS are up to date.
+* Replaced `YYImage` with `SDWebImageWebPCoder` as `YYImage` no longer maintain to support ARM.
+
 ## 0.1.0-nullsafety.0
 
 * Null safety has been added.